1. Static Report 대부분 완료.  > hourly weelky 등등 완료. 그런데 energy prodction부분 버그 잇음.

+++ availability-server/src/main/java/com/unison/scada/availability/api/availability/entity/AvailabilityData.java
@@ -24,7 +24,7 @@ public class AvailabilityData {
 
     @Setter
     @Column(nullable = false)
-    private int time;
+    private Double time;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "availability_type_uuid", nullable = true)
