+++ availability-server/src/main/java/com/unison/scada/availability/api/windfarm/daily/DailyWindFarmDTO.java
@@ -92,6 +92,6 @@ public class DailyWindFarmDTO {
     @Builder
     public static class Availability {
         private final String name;
-        private final int time;
+        private final double time;
     }
 }
